+++ flake.nix
@@ -2,7 +2,7 @@
   description = "T — A Functional Language for Tabular Data";
 
   inputs = {
-    nixpkgs.url = "github:rstats-on-nix/nixpkgs/2026-06-23";
+    nixpkgs.url = "github:rstats-on-nix/nixpkgs/2026-09-23";
     flake-utils.url = "github:numtide/flake-utils";
   };